perf: parallel hy2 UDP demux + wire tuic brutal congestion control
hysteria2:
- demux UDP datagrams with up to GOMAXPROCS (cap 8) receive goroutines;
  quic-go's datagramQueue.Receive is safe for concurrent consumers, so a
  single slow handler no longer caps aggregate packet rate at one core
- closeCleanup via sync.Once (multi-run exit no longer double-closes done)
- closeLocked takes receiveMu around Defragger.Close (fixes pre-existing
  race with concurrent deliverMessage/ReadFrom feeds)
- ReceiveMessage exits on connection context error for temporary errors,
  fixing a stateless-reset tight spin loop
- feed releases pooled datagrams for unknown sessions

tuic:
- parse/export cwnd from share links and wire it through Header.Feature2
  into ClientOption.CWND; SetCongestionController selects brutal with the
  cwnd as target bandwidth (bytes/s, community convention), falling back
  to BBR when unset
- drop the dead CWND:10 hardcode (tuic) and the same in juicity

Regression tests: parallel demux delivery + cleanup, tuic URL cwnd round-trip.

+++ protocol/tuic/client.go
@@ -42,7 +42,9 @@ type ClientOption struct {
 	MaxUdpRelayPacketSize int
 	CongestionController  string
 	ReduceRtt             bool
-	CWND                  int
+	// CWND carries the brutal controller's target bandwidth (bytes per
+	// second); ignored by other controllers.
+	CWND uint64
 }
 
 type clientImpl struct {
